Internalized some methods, Updated example script (See desc.)
Internalized some methods that weren't meant to be exposed.
Updated the example script to use the Answered event handler and changed
the Active property on the contact  to true to minimize confusion.
Added a method to end the active call.
Fixed a mistake with the In- Call UI showing the wrong display image.
Deprecated the Selected event.
Updated some variables with more concise names.